% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_linear_model.R
\name{fit_linear_model}
\alias{fit_linear_model}
\title{Fit a generalized linear model for differential expression}
\usage{
fit_linear_model(
  input_matrix,
  model_matrix,
  size_factors = TRUE,
  group_matrix = NULL,
  gene_specific_model_tensor = NULL,
  kernel_input = NULL,
  gene_names = NULL,
  cell_names = NULL,
  variance = "VI_Estimate",
  inference_method = "SVI",
  method_specific_args = list()
)
}
\arguments{
\item{input_matrix}{Matrix of counts representing gene expression
data for individual cells. Each row corresponds to a gene, and each
column represents a single cell.}

\item{model_matrix}{Matrix also known as design matrix, it represents
the relationship between the response variable and the predictor
variables in the model. Each row represents a cell and each columns
represent a predictor variable (e.g experimental conditions, biological
factors, treatment groups, batch effects, ...)}

\item{size_factors}{Boolean. Decides if a scaling factor for the
expression of each cell should be computed}

\item{group_matrix}{.}

\item{gene_specific_model_tensor}{.}

\item{kernel_input}{.}

\item{gene_names}{Vector containing the names of the genes}

\item{cell_names}{Vector containing the names of the cells}

\item{variance}{String. Either "VI_Estimate" or "Hessian".}

\item{inference_method}{String. Either "SVI" or "HMC"}

\item{method_specific_args}{List containing additional arguments.
The available arguments differs between the inference algorithms.

SVI only:
\itemize{
\item \code{optimizer_name} optimizer, one of "ClippedAdam", "Adam", and "SGD";
\item \code{steps} number of iterations of the optimization algorithm;
\item \code{lr} learning rate for the optimize;
\item \code{gamma_lr} parameters to tune the decay of the learning rate using "ClippedAdam";
\item \code{batch_size} number of data points or observations sampled from the input matrix in
each iteration of the optimization algorithm;
\item \code{threshold} parameters to stop the inference earlier when convergence is reached.
Default value is set to 0, i.e. all steps will be done;
}

HMC only:
\itemize{
\item \code{num_samples} number of iterations after the warmup-phase, it also indicates the
posterior samples each chain will produce;
\item \code{num_chains} number of chains for the optimization algorithm;
\item \code{warmup_steps} number of iterations of the warmup-phase;
}

Shared:
\itemize{
\item \code{cuda} Boolean, indicates if CUDA should be used if available;
\item \code{jit_compile} ;
\item \code{full_cov} ;
\item \code{theta_bounds} ;
\item \code{init_loc} ;
}}
}
\value{
A rdevil object of class \code{rdevil}
}
\description{
Fit a generalized linear model in order to
model the expression of genes coming from a scRNA count matrix.
The resulting object retains the inferred parameters that can
later be tested in order to find differentially expressed genes.
}
