% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdevil-package.R
\docType{package}
\name{rdevil-package}
\alias{rdevil}
\alias{rdevil-package}
\title{rdevil: rdevil}
\description{
An R package to perform fast and scalable single cell differential expression analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://caravagnalab.github.io/rdevil/}
}

}
\author{
\strong{Maintainer}: Giovanni Santacatterina \email{santacatterinagiovanni@gmail.com}

}
\keyword{internal}
