% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_posterior_CI}
\alias{test_posterior_CI}
\title{Test credible interval SVI}
\usage{
test_posterior_CI(inference_res, contrast, credible_mass = 0.95)
}
\arguments{
\item{inference_res}{results coming from \code{fit_linear_model} function}

\item{contrast}{Vector with the same number of rows as the number of
conditions in the design matrix indicating how the coefficients
estimated should be tested against each other.}

\item{credible_mass}{Credible interval required for testing. Default is 0.95.}
}
\value{
A data.frame with \code{gene}, \code{log_FC}, \code{CI_low},
\code{CI_high} and \code{is_zero_in} columns
}
\description{
Test credible interval SVI
}
