% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_posterior_ROPE}
\alias{test_posterior_ROPE}
\title{Test posterior ROPE SVI}
\usage{
test_posterior_ROPE(inference_res, contrast, LFC = 0.5)
}
\arguments{
\item{inference_res}{results coming from \code{fit_linear_model} function}

\item{contrast}{Vector with the same number of rows as the number of
conditions in the design matrix indicating how the coefficients
estimated should be tested against each other.}

\item{LFC}{}
}
\description{
Test posterior ROPE SVI
}
